requirements.txt
discord.py==2.1
asyncpg==0.27.0
